feat(F8): Implement IHM and clean

+++ src/main/java/fr/oc/nico/clambering/service/TopoServiceImpl.java
@@ -10,7 +10,10 @@ import fr.oc.nico.clambering.repository.UtilisateurRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
 import java.util.Date;
+import java.util.List;
 
 @Service("TopoService")
 public class TopoServiceImpl implements TopoService {
@@ -31,14 +34,20 @@ public class TopoServiceImpl implements TopoService {
         String topoLibelle = topoEditForm.getTopoLibelle();
         Region lieu = regionRepository.findByRegionLibelle(topoEditForm.getLieu());
         String description = topoEditForm.getDescription();
-        Date dateParution = topoEditForm.getDateParution();
+        Date dateParution = null;
+        try {
+            dateParution = new SimpleDateFormat("yyyy-MM-dd").parse(topoEditForm.getDateParution());
+        } catch (ParseException e) {
+            e.printStackTrace();
+        }
         Utilisateur proprietaire = utilisateurRepository.findByPseudo(user);
         Topo topo = new Topo(topoLibelle, lieu, description, dateParution, proprietaire);
         topoRepository.save(topo);
     }
 
     @Override
-    public void switchDispo(Topo topo) {
+    public void switchDispo(Integer topoId) {
+        Topo topo = topoRepository.findById(topoId).orElse(null);
         if (topo.getDispo()) {
             topo.setDispo(false);
         } else {
@@ -48,7 +57,22 @@ public class TopoServiceImpl implements TopoService {
     }
 
     @Override
-    public void supprimerTopo(Topo topo) {
-        topoRepository.delete(topo);
+    public void supprimerTopo(String user, Integer topoId) {
+        Topo topo = topoRepository.findById(topoId).orElse(null);
+        String propietaire = topo.getProprietaire().getPseudo();
+        if (propietaire.equals(user)) {
+            topoRepository.delete(topo);
+        }
+    }
+
+    @Override
+    public List<Topo> getMesTopos(String user) {
+        Utilisateur proprietaire = utilisateurRepository.findByPseudo(user);
+        return topoRepository.findAllByProprietaire(proprietaire);
+    }
+
+    @Override
+    public List<Region> getFormData() {
+        return regionRepository.findAll();
     }
 }
